+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 
 # Release Notes
 
+# 0.3.0 / 2018-02-13
+- added msgpack encoding
+- added headers/body separation to events
+
 # 0.2.0 / 2018-02-07
-**Breaking change!**
 - create an interface for the queue
 - move the original implementation to rabbitmq package and keep the interface in the geb package
 - unexport `Timeout` struct field from rabbitmq.Queue, it is set through an Option, so don't need to be exposed